refactor(stage-ui): rename analytics runtime path
