+++ packages/stage-ui/package.json
@@ -31,8 +31,8 @@
     "./features/motions/live2d": "./src/features/motions/live2d/index.ts",
     "./libs/inference/adapters/*": "./src/libs/inference/adapters/*.ts",
     "./libs/inference": "./src/libs/inference/index.ts",
-    "./libs/analytics": "./src/libs/analytics/index.ts",
-    "./libs/analytics/*": "./src/libs/analytics/*.ts",
+    "./libs/product-signals": "./src/libs/product-signals/index.ts",
+    "./libs/product-signals/*": "./src/libs/product-signals/*.ts",
     "./libs/pinia": "./src/libs/pinia/index.ts",
     "./libs/providers/stream-transcription": "./src/libs/providers/stream-transcription/index.ts",
     "./libs/providers/providers/aliyun-nls": "./src/libs/providers/providers/aliyun-nls/index.ts",
