+++ packages/stage-ui/src/libs/product-signals/events/chat/events/message.ts
