+++ packages/stage-ui/src/libs/product-signals/events/controls-island/events/action.ts
