+++ packages/stage-ui/src/libs/product-signals/events/interaction.ts
