+++ packages/stage-ui/src/libs/product-signals/events/mcp/events/server.ts
