+++ packages/stage-ui/src/libs/product-signals/events/mcp/index.ts
