+++ packages/stage-ui/src/libs/product-signals/headers.ts
