+++ packages/stage-ui/src/libs/product-signals/posthog.ts
